Move CLI structs into separate file
